helt fääärdida, typ

+++ clientserver/filesystemdatabase.cc
@@ -18,7 +18,7 @@ FileSystemDatabase::FileSystemDatabase(string root_dir_path) {
 	struct stat statbuf;
 	if (stat(root_dir.c_str(), &statbuf) != -1) { 	//checking status on root_dir
    		if (S_ISDIR(statbuf.st_mode)) {				//checking if it was a directory
-   			cout<<"db finns redan"<<endl;
+   			cout<<"Databasmap finns redan"<<endl;
 			ifstream ng_meta_file;
 			int nextID;
 			ng_meta_file.open(root_dir+"/meta");
@@ -26,55 +26,55 @@ FileSystemDatabase::FileSystemDatabase(string root_dir_path) {
 			Newsgroup::setNextNewsgroupID(nextID);
 		}
 	}else {
+		cout<<"Skapar databasmap"<<endl;
    		string cmd = "mkdir " + root_dir_path;
 		system(cmd.c_str());	
 		cmd = "touch meta";
 		string ng_meta_path = root_dir_path+"/meta";
 		ofstream out_ng_meta_file(ng_meta_path);
-		out_meta_file <<"1";
-		out_meta_file.close(); 
+		out_ng_meta_file <<"1";
+		out_ng_meta_file.close(); 
 	}
 }
 
-void FileSystemDatabase::addNewsgroup(Newsgroup ng){
+int FileSystemDatabase::addNewsgroup(Newsgroup ng){
+	if (findNewsgroup(ng.getTitle()) != -1) {
+		return -1;
+	}	
+
 	string ng_path = root_dir+"/"+to_string(ng.getID());
 	string cmd = "mkdir " + ng_path;
 	system(cmd.c_str());
 	//increment next newsgroup id in metafile.
-	string ng_meta_path = root_dir_path+"/meta";
+	string ng_meta_path = root_dir+"/meta";
 	ifstream ng_meta_file(ng_meta_path);
 	string nextID;
 	ng_meta_file >> nextID;
-	ng_meta_file.close()
+	ng_meta_file.close();
 	ofstream out_ng_meta_file(ng_meta_path);
-	out_ng_meta_file << nextID+1;
-	out_ng_meta_file.close()
+	out_ng_meta_file << stoi(nextID)+1;
+	out_ng_meta_file.close();
 
 	//creating meta_file for ng
 
 	ofstream meta_file(ng_path+"/meta");
-	meta_file<<"0"<<endl<<ng.getTitle()<<endl;
+	meta_file<<"1"<<endl<<ng.getTitle()<<endl;
 	meta_file.close();
 
-	for(auto a : *ng.getArticles()){ //vad e detta?? lägger vi in artiklar från en ng som vi precis skapat??
-		Article article = a.second;
-		addArticle(ng.getID(), article.getTitle(), article.getAuthor(), article.getText());
-	}
+	return ng.getID();
 }
 
 int FileSystemDatabase::removeNewsgroup(int id){
-	string ng_path = root_dir+"/"+to_string(id);
-	
-	//kan l'gga till h'r så att om vi tar bort det högsta idt så updaterar vi nästa id? men då måste vi
-	//räkna alla ifall de tidigare också tagits bort. kanske är oödigt??
-
-	string cmd = "rm -rf " + ng_path;
-	system(cmd.c_str());
-
-	return id; //todo: wät?
+	int t = findNewsgroup(id);
+	if(t != -1){
+		string ng_path = root_dir+"/"+to_string(id);
+		string cmd = "rm -rf " + ng_path;
+		system(cmd.c_str());
+	}
+	return t;
 }
 
-std::vector<Newsgroup> FileSystemDatabase::listNewsgroups(){
+vector<Newsgroup> FileSystemDatabase::listNewsgroups(){
 	vector<Newsgroup> ngs;
 	vector<int> ngIDs = readDirectory(root_dir);
 	ifstream meta_file;
@@ -82,8 +82,8 @@ std::vector<Newsgroup> FileSystemDatabase::listNewsgroups(){
 	string title;
 	for(auto ngID : ngIDs){
 		meta_file.open(root_dir+"/"+to_string(ngID)+"/meta");
-		meta_file >> trash;
-		meta_file >> title;
+		getline(meta_file,trash);
+		getline(meta_file,title);
 		ngs.push_back(Newsgroup(title, ngID));
 		meta_file.close();
 		title = "";
@@ -94,9 +94,18 @@ std::vector<Newsgroup> FileSystemDatabase::listNewsgroups(){
 /* Returns -1 if name is not found. */
 int FileSystemDatabase::findNewsgroup(const std::string &name){
 	vector<Newsgroup> ngs = listNewsgroups();
-	auto found = find_if(ngs.begin(), ngs.end(), [&](Newsgroup n){return n.getTitle() == name;});
+	auto found = find_if(ngs.begin(), ngs.end(), [&](const Newsgroup &n){ return n.getTitle() == name;});
 	if(found != ngs.end()){
 		return found->getID();
+	}
+	return -1;
+}
+
+int FileSystemDatabase::findNewsgroup(int id){
+	vector<Newsgroup> ngs = listNewsgroups();
+	auto found = find_if(ngs.begin(), ngs.end(), [id](const Newsgroup &n){ return n.getID() == id;});
+	if(found != ngs.end()){
+		return id;
 	}
 	return -1;
 }
@@ -108,10 +117,10 @@ int FileSystemDatabase::getAndIncrementNextArticleId(int ngID){
 
 	ifstream meta_file(meta_path);
 	string id;
-	meta_file >> id;
+	getline(meta_file,id);
 	nextID = stoi(id);
 	string title;
-	meta_file >> title;
+	getline(meta_file,title);
 	meta_file.close();
 
 	ofstream out_meta_file(meta_path);
@@ -121,22 +130,28 @@ int FileSystemDatabase::getAndIncrementNextArticleId(int ngID){
 }
 
 int FileSystemDatabase::addArticle(int ngID, const std::string &title, const std::string &author, const std::string &text){
-	int aID = getAndIncrementNextArticleId(ngID);
-
-	string a_path = root_dir + "/" + to_string(ngID) + "/"+to_string(aID);
-	ofstream a_file(a_path);
-	a_file<<title<<endl<<author<<endl<<text<<endl;
-	a_file.close();
-	return ngID;
+	if(findNewsgroup(ngID)!= -1){
+		string ng_path = root_dir + "/" + to_string(ngID);
+		int aID = getAndIncrementNextArticleId(ngID);
+		string a_path = ng_path + "/"+to_string(aID);
+		ofstream a_file(a_path);
+		a_file<<title<<endl<<author<<endl<<text<<endl;
+		a_file.close();
+		return ngID;
+	}
+	return -1;
 }
 
 int FileSystemDatabase::listArticles(int ngID, std::vector<std::pair<int ,Article>> &vec){
-	vector<int> aids = readDirectory(root_dir+"/"+to_string(ngID));
-	int check;
-	for(auto aID : aids)
-		vec.push_back(make_pair(aID, getArticle(ngID, aID, check)));
-	//todo: check?
-	return ngID;
+	if(findNewsgroup(ngID)!=-1){
+		string ng_path = root_dir + "/" + to_string(ngID);
+		vector<int> aids = readDirectory(root_dir+"/"+to_string(ngID));
+		int check;
+		for(auto aID : aids)
+			vec.push_back(make_pair(aID, getArticle(ngID, aID, check)));
+		return ngID;
+	}
+	return -1;
 }
 
 vector<int> FileSystemDatabase::readDirectory(string dir_path){
@@ -154,24 +169,39 @@ vector<int> FileSystemDatabase::readDirectory(string dir_path){
 }
 
 int FileSystemDatabase::deleteArticle(int ngID, int artID){
-	string cmd = "rm " + root_dir + "/" + to_string(ngID) + "/" + to_string(artID);
+	int check;
+	getArticle(ngID, artID, check);
+	if (check == -1)
+		return -1;
+	if (check == -2)
+		return -2;
+	string ng_path = root_dir + "/" + to_string(ngID);
+	string cmd = "rm " + ng_path + "/" + to_string(artID);
 	system(cmd.c_str());
-	//todo: check;
-	return artID; //todo: wät?
+	//todo: check?
+	return artID;
 }
 
 Article FileSystemDatabase::getArticle(int ngID, int artID, int &check){
-	string file_path = root_dir+"/"+to_string(ngID)+"/"+to_string(artID);
-	ifstream a_stream;
-	a_stream.open(file_path.c_str());
-	if(!a_stream){
-		check = -1;
-		return Article();
+	if (findNewsgroup(ngID) != -1) {
+		string file_path = root_dir+"/"+to_string(ngID)+"/"+to_string(artID);
+		ifstream a_stream;
+		a_stream.open(file_path.c_str());
+		if(!a_stream){
+			check = -2;
+			return Article();
+		}
+		string title, author, text;
+		getline(a_stream,title);
+		getline(a_stream,author);
+		string line;
+		while(getline(a_stream,line)){
+			text += line+"\n";
+		}
+		text.pop_back();
+		check = ngID;
+		return Article(title, author, text);
 	}
-	string title, author, text;
-	a_stream >> title;
-	a_stream >> author;
-	a_stream >> text;
-	check = ngID;
-	return Article(title, author, text);
+	check = -1;
+	return Article();
 }
